pyproject.toml
[build-system]
requires = ["hatchling>=1.21.0"]
build-backend = "hatchling.build"

[project]
name = "meq-bench"
version = "2.0.0"
description = "A Resource-Efficient Benchmark for Evaluating Audience-Adaptive Explanation Quality in Medical LLMs"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "MEQ-Bench Team", email = "contact@meq-bench.org" }
]
maintainers = [
    { name = "MEQ-Bench Team", email = "contact@meq-bench.org" }
]
keywords = [
    "medical",
    "nlp",
    "benchmark",
    "evaluation",
    "llm",
    "ai",
    "healthcare",
    "machine-learning",
    "audience-adaptive",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "Intended Audience :: Healthcare Industry",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3 :: Only",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Medical Science Apps.",
    "Typing :: Typed",
]
dependencies = [
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "PyYAML>=6.0.1",
    "requests>=2.31.0",
    "textstat>=0.7.3",
    "httpx>=0.27.0",
    "structlog>=24.1.0",
    "tenacity>=8.2.0",
    "rich>=13.7.0",
    "typer>=0.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "pytest-asyncio>=0.23.0",
    "pytest-xdist>=3.5.0",
    "pytest-benchmark>=4.0.0",
    "ruff>=0.4.0",
    "mypy>=1.9.0",
    "pre-commit>=3.6.0",
    "bandit[toml]>=1.7.8",
    "types-PyYAML>=6.0.12",
    "types-requests>=2.31.0",
    "respx>=0.21.0",
]
test = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "pytest-asyncio>=0.23.0",
    "pytest-xdist>=3.5.0",
    "respx>=0.21.0",
    "hypothesis>=6.100.0",
]
docs = [
    "mkdocs>=1.5.0",
    "mkdocs-material>=9.5.0",
    "mkdocstrings[python]>=0.24.0",
    "mkdocs-gen-files>=0.5.0",
    "mkdocs-literate-nav>=0.6.0",
]
ml = [
    "torch>=2.2.0",
    "transformers>=4.38.0",
    "datasets>=2.17.0",
    "scikit-learn>=1.4.0",
    "sentence-transformers>=2.5.0",
    "bert-score>=0.3.13",
    "spacy>=3.7.0",
    "nltk>=3.8.1",
]
llm = [
    "openai>=1.12.0",
    "anthropic>=0.18.0",
    "google-generativeai>=0.4.0",
]
apple = [
    "mlx>=0.10.0",
    "mlx-lm>=0.10.0",
]
analysis = [
    "jupyter>=1.0.0",
    "matplotlib>=3.8.0",
    "seaborn>=0.13.0",
    "plotly>=5.19.0",
    "scipy>=1.12.0",
    "tqdm>=4.66.0",
]
all = [
    "meq-bench[dev,test,docs,ml,llm,analysis]",
]

[project.urls]
Homepage = "https://github.com/heilcheng/MEQ-Bench"
Documentation = "https://meq-bench.readthedocs.io"
Repository = "https://github.com/heilcheng/MEQ-Bench.git"
Issues = "https://github.com/heilcheng/MEQ-Bench/issues"
Changelog = "https://github.com/heilcheng/MEQ-Bench/blob/main/CHANGELOG.md"

[project.scripts]
meq-bench = "src.cli:app"
meq-evaluate = "src.evaluator:main"

[tool.hatch.build.targets.sdist]
include = [
    "/src",
    "/tests",
    "/data",
]
exclude = [
    "/.github",
    "/docs",
    "/.venv",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]

# Ruff - Fast Python linter and formatter (replaces flake8, isort, black)
[tool.ruff]
target-version = "py310"
line-length = 120
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
    "TCH",    # flake8-type-checking
    "PTH",    # flake8-use-pathlib
    "ERA",    # eradicate (commented code)
    "PL",     # Pylint
    "RUF",    # Ruff-specific rules
    "PERF",   # Perflint
    "FURB",   # refurb
    "LOG",    # flake8-logging
    "G",      # flake8-logging-format
]
ignore = [
    "PLR0913",  # Too many arguments
    "PLR2004",  # Magic value comparison
    "PLR0912",  # Too many branches
    "E501",     # Line too long (handled by formatter)
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["ARG", "PLR2004", "S101"]
"__init__.py" = ["F401"]

[tool.ruff.lint.isort]
known-first-party = ["src"]
force-single-line = false
lines-after-imports = 2

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
docstring-code-format = true

# MyPy configuration
[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_configs = true
show_error_codes = true
show_column_numbers = true
pretty = true
plugins = ["pydantic.mypy"]

[[tool.mypy.overrides]]
module = [
    "textstat.*",
    "bert_score.*",
    "sentence_transformers.*",
    "transformers.*",
    "torch.*",
    "sklearn.*",
    "spacy.*",
    "nltk.*",
    "mlx.*",
    "mlx_lm.*",
]
ignore_missing_imports = true

[tool.pydantic-mypy]
init_forbid_extra = true
init_typed = true
warn_required_dynamic_aliases = true

# Pytest configuration
[tool.pytest.ini_options]
minversion = "8.0"
testpaths = ["tests"]
pythonpath = ["."]
addopts = [
    "-ra",
    "-q",
    "--strict-markers",
    "--strict-config",
    "-p", "no:warnings",
    "--tb=short",
]
markers = [
    "unit: Unit tests (fast, isolated)",
    "integration: Integration tests (may require external services)",
    "slow: Slow tests (>5s)",
    "requires_gpu: Tests requiring GPU",
    "requires_api: Tests requiring API keys",
]
asyncio_mode = "auto"
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::PendingDeprecationWarning",
]

# Coverage configuration
[tool.coverage.run]
source = ["src"]
branch = true
parallel = true
omit = [
    "*/tests/*",
    "*/__init__.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if TYPE_CHECKING:",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "@abstractmethod",
    "@overload",
]
fail_under = 80
show_missing = true
skip_covered = true

[tool.coverage.html]
directory = "htmlcov"

# Bandit security scanning
[tool.bandit]
exclude_dirs = ["tests", "docs"]
skips = ["B101", "B601"]  # Skip assert and shell injection for subprocess

# Semantic versioning with commitizen
[tool.commitizen]
name = "cz_conventional_commits"
tag_format = "v$version"
version_scheme = "semver"
version_provider = "pep621"
update_changelog_on_bump = true
major_version_zero = true
